move log dumping into platform actions (#12677)
* move log dumping into platform actions

* move more actions into platform specific. split out avatar stuff (#12686)
