--- shared/logger/dump-log-fs.native.js
@@ -1,3 +1,0 @@
-// @flow
-const dumpLogs = () => {}
-export default dumpLogs
